reorganized subagent-based tools
