+++ mdagent/tools/__init__.py
@@ -1,7 +1,3 @@
-from .maketools import make_tools
-from .base_tools.registry.path_registry import PathRegistry
-
-from.base_tools.registry.registry_tools import ListRegistryPaths, MapPath2Name
 from .base_tools.clean_tools import (
     AddHydrogensCleaningTool,
     CleaningTools,
@@ -10,6 +6,8 @@ from .base_tools.clean_tools import (
 )
 from .base_tools.md_util_tools import Name2PDBTool, get_pdb
 from .base_tools.plot_tools import SimulationOutputFigures
+from .base_tools.registry.path_registry import PathRegistry
+from .base_tools.registry.registry_tools import ListRegistryPaths, MapPath2Name
 from .base_tools.search_tools import Scholar2ResultLLM
 from .base_tools.setup_and_run import SetUpAndRunTool, SimulationFunctions
 from .base_tools.vis_tools import (
@@ -18,6 +16,8 @@ from .base_tools.vis_tools import (
     VisFunctions,
     VisualizationToolRender,
 )
+from .maketools import make_tools
+from .subagent_tools import GetNewTool, SkillQuery, SkillUpdate
 
 __all__ = [
     "Scholar2ResultLLM",
@@ -37,8 +37,8 @@ __all__ = [
     "CheckDirectoryFiles",
     "PlanBVisualizationTool",
     "SimulationOutputFigures",
-
     "make_tools",
-    "PathRegistry",
+    "GetNewTool",
+    "SkillQuery",
+    "SkillUpdate",
 ]
-
